changes done with put method

+++ Backend/S_Auth/djpr_Registery/djap_Settings/views.py
@@ -1,19 +1,20 @@
 from django.shortcuts import render
 from djap_login.models import FullUser
 import requests
+import json
 from django.shortcuts import render
 from django.http import HttpResponse, JsonResponse
 
 def delete(request):
-	if (request.method == 'POST'):
-		username = request.POST.get('username')
+	if (request.method == 'DELETE'):
+		username = request.DELETE.get('username')
 
-		if not username:
-			return JsonResponse({"error": "Missing credentials"}, status=400)
+		if not username or not isinstance(username, str):
+			return JsonResponse({"error": "Missing datas"}, status=400)
 
 		user = FullUser.objects.filter(username=username).first()
 		if (user == None):
-			return JsonResponse({"error": "Invalid Credentials"}, status=401)
+			return JsonResponse({"error": "Invalid datas"}, status=400)
 
 		try:
 			user.delete()
@@ -24,16 +25,19 @@ def delete(request):
 			return JsonResponse({"error": f"Unexpected error: {str(e)}"}, status=500)
 
 def choose_lang(request):
-	if (request.method == 'POST'):
-		username = request.POST.get('username')
-		new_lang = request.POST.get('newLang')
+	if (request.method == 'PUT'):
+		data = json.loads(request.body)
+		username = data.get('username')
+		new_lang = data.get('newLang')
 
 		if not username or not new_lang:
-			return JsonResponse({"error": "Missing credentials"}, status=400)
+			return JsonResponse({"error": "Missing datas"}, status=400)
 
 		user = FullUser.objects.filter(username=username).first()
+		if (not isinstance(username, str) or not isinstance(new_lang, str)):
+			return JsonResponse({"error": "Invalid datas"}, status=400)
 		if (user == None):
-			return JsonResponse({"error": "Invalid Credentials"}, status=401)
+			return JsonResponse({"error": "Invalid datas"}, status=400)
 
 		try:
 			user.language = new_lang
@@ -45,20 +49,20 @@ def choose_lang(request):
 			return JsonResponse({"error": f"Unexpected error: {str(e)}"}, status=500)
 
 def choose_verif(request):
-	if (request.method == 'POST'):
-		username = request.POST.get('username')
-		new_2fa = request.POST.get('new2fa')
+	if (request.method == 'PUT'):
+		data = json.loads(request.body)
+		username = data.get('username')
+		new_2fa = data.get('new2fa')
 
 		if not username or not new_2fa:
-			return JsonResponse({"error": "Missing credentials"}, status=400)
+			return JsonResponse({"error": "Missing datas"}, status=400)
 		
 		if isinstance(new_2fa, str) and (new_2fa.lower() == 'true' or new_2fa.lower() == 'false'):
 			new_2fa = new_2fa.lower() == 'true'
 		elif isinstance(new_2fa, bool):
 			pass
 		else:
-			return JsonResponse({"error": "Wrong credentials"}, status=400)
-
+			return JsonResponse({"error": "Wrong datas"}, status=400)
 		user = FullUser.objects.filter(username=username).first()
 		if (user == None):
 			return JsonResponse({"error": "Invalid Credentials"}, status=401)
@@ -76,12 +80,12 @@ def get_lang(request):
 	if (request.method == 'GET'):
 		username = request.GET.get('username')
 
-		if not username:
-			return JsonResponse({"error": "Missing credentials"}, status=400)
+		if not username or not isinstance(username, str):
+			return JsonResponse({"error": "Missing datas"}, status=400)
 
 		user = FullUser.objects.filter(username=username).first()
 		if (user == None):
-			return JsonResponse({"error": "Invalid Credentials"}, status=401)
+			return JsonResponse({"error": "Invalid datas"}, status=400)
 
 		try:
 			res = user.language
@@ -95,12 +99,12 @@ def get_verif(request):
 	if (request.method == 'GET'):
 		username = request.GET.get('username')
 
-		if not username:
-			return JsonResponse({"error": "Missing credentials"}, status=400)
+		if not username or not isinstance(username, str):
+			return JsonResponse({"error": "Missing datas"}, status=400)
 
 		user = FullUser.objects.filter(username=username).first()
 		if (user == None):
-			return JsonResponse({"error": "Invalid Credentials"}, status=401)
+			return JsonResponse({"error": "Invalid datas"}, status=400)
 
 		try:
 			res = user.secu
